refactor(gateway): extract memory and subagent WS handlers (#818)
