+++ packages/gateway/src/ws/protocol/handler.ts
@@ -33,33 +33,7 @@ import {
   WsCapabilityReadyRequest,
   WsAttemptEvidenceRequest,
   WsMessageEnvelope,
-  WsMemorySearchRequest,
-  WsMemorySearchResult,
-  WsMemoryListRequest,
-  WsMemoryListResult,
-  WsMemoryGetRequest,
-  WsMemoryGetResult,
-  WsMemoryCreateRequest,
-  WsMemoryCreateResult,
-  WsMemoryUpdateRequest,
-  WsMemoryUpdateResult,
-  WsMemoryDeleteRequest,
-  WsMemoryDeleteResult,
-  WsMemoryForgetRequest,
-  WsMemoryForgetResult,
-  WsMemoryExportRequest,
-  WsMemoryExportResult,
   WsTaskExecuteResult,
-  WsSubagentSpawnRequest,
-  WsSubagentSpawnResult,
-  WsSubagentListRequest,
-  WsSubagentListResult,
-  WsSubagentGetRequest,
-  WsSubagentGetResult,
-  WsSubagentSendRequest,
-  WsSubagentSendResult,
-  WsSubagentCloseRequest,
-  WsSubagentCloseResult,
   clientCapabilityFromDescriptorId,
   parseTyrumKey,
 } from "@tyrum/schemas";
@@ -71,74 +45,17 @@ import type {
   WsResponseEnvelope,
 } from "@tyrum/schemas";
 import type { ConnectedClient } from "../connection-manager.js";
-import type { WsBroadcastAudience } from "../audience.js";
-import { WORKBOARD_WS_AUDIENCE } from "../workboard-audience.js";
 import { emitPairingApprovedEvent } from "../pairing-approved.js";
 import { toApprovalContract } from "../../modules/approval/to-contract.js";
 import { executeCommand } from "../../modules/commands/dispatcher.js";
 import { hasAnyRequiredScope } from "../../modules/auth/scopes.js";
 import { resolveWsRequestRequiredScopes } from "../../modules/authz/ws-scope-matrix.js";
 import { isSafeSuggestedOverridePattern } from "../../modules/policy/override-guardrails.js";
-import { WorkboardDal } from "../../modules/workboard/dal.js";
 import type { ProtocolDeps } from "./types.js";
-import { broadcastEvent, errorResponse, workboardErrorResponse } from "./helpers.js";
+import { broadcastEvent, errorResponse } from "./helpers.js";
 import { handleWorkboardMessage } from "./workboard-handlers.js";
-
-const OPERATOR_MEMORY_EVENT_AUDIENCE = {
-  roles: ["client"],
-  required_scopes: ["operator.read"],
-} as const satisfies WsBroadcastAudience;
-
-async function maybeRunMemoryV1BudgetConsolidation(params: {
-  deps: ProtocolDeps;
-  op: "create" | "update";
-  agentId: string;
-  memoryItemId: string;
-}): Promise<void> {
-  if (!params.deps.memoryV1BudgetsProvider) return;
-  if (!params.deps.memoryV1Dal) return;
-  try {
-    const budgets = await params.deps.memoryV1BudgetsProvider(params.agentId);
-    const consolidation = await params.deps.memoryV1Dal.consolidateToBudgets({
-      budgets,
-      agentId: params.agentId,
-    });
-
-    for (const created of consolidation.created_items) {
-      broadcastEvent(
-        {
-          event_id: crypto.randomUUID(),
-          type: "memory.item.created",
-          occurred_at: created.created_at,
-          payload: { item: created },
-        },
-        params.deps,
-        OPERATOR_MEMORY_EVENT_AUDIENCE,
-      );
-    }
-
-    for (const tombstone of consolidation.deleted_tombstones) {
-      broadcastEvent(
-        {
-          event_id: crypto.randomUUID(),
-          type: "memory.item.deleted",
-          occurred_at: tombstone.deleted_at,
-          payload: { tombstone },
-        },
-        params.deps,
-        OPERATOR_MEMORY_EVENT_AUDIENCE,
-      );
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    params.deps.logger?.error("memory.v1.consolidation_failed", {
-      op: params.op,
-      agent_id: params.agentId,
-      memory_item_id: params.memoryItemId,
-      error: message,
-    });
-  }
-}
+import { handleSubagentMessage } from "./subagent-handlers.js";
+import { handleMemoryMessage } from "./memory-handlers.js";
 
 // ---------------------------------------------------------------------------
 // Client message handling
@@ -1049,368 +966,8 @@ export async function handleClientMessage(
     return { request_id: msg.request_id, type: msg.type, ok: true, result };
   }
 
-  if (msg.type === "subagent.spawn") {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may spawn subagents",
-      );
-    }
-    if (!deps.db) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "subagent.spawn not supported",
-      );
-    }
-
-    const parsedReq = WsSubagentSpawnRequest.safeParse(msg);
-    if (!parsedReq.success) {
-      return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-        issues: parsedReq.error.issues,
-      });
-    }
-
-    const dal = new WorkboardDal(deps.db);
-    try {
-      const payload = parsedReq.data.payload;
-      const subagentId = crypto.randomUUID();
-      const sessionKey = `agent:${payload.agent_id}:subagent:${subagentId}`;
-      const subagent = await dal.createSubagent({
-        scope: payload,
-        subagent: {
-          execution_profile: payload.execution_profile,
-          session_key: sessionKey,
-          work_item_id: payload.work_item_id,
-          work_item_task_id: payload.work_item_task_id,
-          lane: "subagent",
-          status: "running",
-        },
-        subagentId,
-      });
-
-      broadcastEvent(
-        {
-          event_id: crypto.randomUUID(),
-          type: "subagent.spawned",
-          occurred_at: new Date().toISOString(),
-          scope: { kind: "agent", agent_id: subagent.agent_id },
-          payload: { subagent },
-        },
-        deps,
-        WORKBOARD_WS_AUDIENCE,
-      );
-
-      const result = WsSubagentSpawnResult.parse({ subagent });
-      return { request_id: msg.request_id, type: msg.type, ok: true, result };
-    } catch (err) {
-      return workboardErrorResponse(msg.request_id, msg.type, err, deps);
-    }
-  }
-
-  if (msg.type === "subagent.list") {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may list subagents",
-      );
-    }
-    if (!deps.db) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "subagent.list not supported",
-      );
-    }
-
-    const parsedReq = WsSubagentListRequest.safeParse(msg);
-    if (!parsedReq.success) {
-      return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-        issues: parsedReq.error.issues,
-      });
-    }
-
-    const dal = new WorkboardDal(deps.db);
-    try {
-      const payload = parsedReq.data.payload;
-      const { subagents, next_cursor } = await dal.listSubagents({
-        scope: payload,
-        statuses: payload.statuses,
-        limit: payload.limit,
-        cursor: payload.cursor,
-      });
-      const result = WsSubagentListResult.parse({ subagents, next_cursor });
-      return { request_id: msg.request_id, type: msg.type, ok: true, result };
-    } catch (err) {
-      return workboardErrorResponse(msg.request_id, msg.type, err, deps);
-    }
-  }
-
-  if (msg.type === "subagent.get") {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may fetch subagents",
-      );
-    }
-    if (!deps.db) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "subagent.get not supported",
-      );
-    }
-
-    const parsedReq = WsSubagentGetRequest.safeParse(msg);
-    if (!parsedReq.success) {
-      return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-        issues: parsedReq.error.issues,
-      });
-    }
-
-    const dal = new WorkboardDal(deps.db);
-    try {
-      const payload = parsedReq.data.payload;
-      const subagent = await dal.getSubagent({ scope: payload, subagent_id: payload.subagent_id });
-      if (!subagent) {
-        return errorResponse(msg.request_id, msg.type, "not_found", "subagent not found");
-      }
-      const result = WsSubagentGetResult.parse({ subagent });
-      return { request_id: msg.request_id, type: msg.type, ok: true, result };
-    } catch (err) {
-      return workboardErrorResponse(msg.request_id, msg.type, err, deps);
-    }
-  }
-
-  if (msg.type === "subagent.send") {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may send subagent messages",
-      );
-    }
-    if (!deps.db || !deps.agents) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "subagent.send not supported",
-      );
-    }
-
-    const parsedReq = WsSubagentSendRequest.safeParse(msg);
-    if (!parsedReq.success) {
-      return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-        issues: parsedReq.error.issues,
-      });
-    }
-
-    const dal = new WorkboardDal(deps.db);
-    const payload = parsedReq.data.payload;
-    let subagent: Awaited<ReturnType<WorkboardDal["getSubagent"]>>;
-    try {
-      subagent = await dal.getSubagent({ scope: payload, subagent_id: payload.subagent_id });
-    } catch (err) {
-      return workboardErrorResponse(msg.request_id, msg.type, err, deps);
-    }
-
-    if (!subagent) {
-      return errorResponse(msg.request_id, msg.type, "not_found", "subagent not found");
-    }
-    if (subagent.status !== "running") {
-      return errorResponse(msg.request_id, msg.type, "invalid_state", "subagent is not running");
-    }
-
-    void (async () => {
-      try {
-        const runtime = await deps.agents!.getRuntime(payload.agent_id);
-        const res = await runtime.turn({
-          channel: "subagent",
-          thread_id: subagent.subagent_id,
-          message: payload.content,
-          metadata: {
-            tyrum_key: subagent.session_key,
-            lane: subagent.lane,
-            subagent_id: subagent.subagent_id,
-            ...(subagent.work_item_id ? { work_item_id: subagent.work_item_id } : {}),
-            ...(subagent.work_item_task_id
-              ? { work_item_task_id: subagent.work_item_task_id }
-              : {}),
-          },
-        });
-
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "subagent.output",
-            occurred_at: new Date().toISOString(),
-            scope: { kind: "agent", agent_id: subagent.agent_id },
-            payload: {
-              tenant_id: payload.tenant_id,
-              agent_id: payload.agent_id,
-              workspace_id: payload.workspace_id,
-              subagent_id: subagent.subagent_id,
-              ...(subagent.work_item_id ? { work_item_id: subagent.work_item_id } : {}),
-              ...(subagent.work_item_task_id
-                ? { work_item_task_id: subagent.work_item_task_id }
-                : {}),
-              kind: "final",
-              content: res.reply ?? "",
-            },
-          },
-          deps,
-          WORKBOARD_WS_AUDIENCE,
-        );
-      } catch (err) {
-        const message = err instanceof Error ? err.message : String(err);
-        deps.logger?.warn("ws.subagent_send_failed", {
-          request_id: msg.request_id,
-          client_id: client.id,
-          request_type: msg.type,
-          subagent_id: payload.subagent_id,
-          error: message,
-        });
-
-        try {
-          const failed = await dal.markSubagentFailed({
-            scope: payload,
-            subagent_id: payload.subagent_id,
-            reason: message,
-          });
-          if (failed) {
-            broadcastEvent(
-              {
-                event_id: crypto.randomUUID(),
-                type: "subagent.updated",
-                occurred_at: new Date().toISOString(),
-                scope: { kind: "agent", agent_id: failed.agent_id },
-                payload: { subagent: failed },
-              },
-              deps,
-              WORKBOARD_WS_AUDIENCE,
-            );
-          }
-        } catch (err) {
-          const updateMessage = err instanceof Error ? err.message : String(err);
-          deps.logger?.warn("ws.subagent_failure_update_failed", {
-            request_id: msg.request_id,
-            client_id: client.id,
-            request_type: msg.type,
-            subagent_id: payload.subagent_id,
-            error: updateMessage,
-          });
-        }
-      }
-    })();
-
-    const result = WsSubagentSendResult.parse({ accepted: true });
-    return { request_id: msg.request_id, type: msg.type, ok: true, result };
-  }
-
-  if (msg.type === "subagent.close") {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may close subagents",
-      );
-    }
-    if (!deps.db) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "subagent.close not supported",
-      );
-    }
-
-    const parsedReq = WsSubagentCloseRequest.safeParse(msg);
-    if (!parsedReq.success) {
-      return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-        issues: parsedReq.error.issues,
-      });
-    }
-
-    const dal = new WorkboardDal(deps.db);
-    try {
-      const payload = parsedReq.data.payload;
-      const closing = await dal.closeSubagent({
-        scope: payload,
-        subagent_id: payload.subagent_id,
-        reason: payload.reason,
-      });
-      if (!closing) {
-        return errorResponse(msg.request_id, msg.type, "not_found", "subagent not found");
-      }
-
-      // Failed/closed subagents are terminal; treat close as idempotent no-op (no lifecycle events).
-      if (closing.status === "failed" || closing.status === "closed") {
-        const result = WsSubagentCloseResult.parse({ subagent: closing });
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      broadcastEvent(
-        {
-          event_id: crypto.randomUUID(),
-          type: "subagent.updated",
-          occurred_at: new Date().toISOString(),
-          scope: { kind: "agent", agent_id: closing.agent_id },
-          payload: { subagent: closing },
-        },
-        deps,
-        WORKBOARD_WS_AUDIENCE,
-      );
-
-      const closed = await dal.markSubagentClosed({
-        scope: payload,
-        subagent_id: payload.subagent_id,
-      });
-      const finalized = closed ?? closing;
-
-      if (finalized.status === "closed") {
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "subagent.closed",
-            occurred_at: new Date().toISOString(),
-            scope: { kind: "agent", agent_id: finalized.agent_id },
-            payload: { subagent: finalized },
-          },
-          deps,
-          WORKBOARD_WS_AUDIENCE,
-        );
-      } else if (finalized.status !== closing.status) {
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "subagent.updated",
-            occurred_at: new Date().toISOString(),
-            scope: { kind: "agent", agent_id: finalized.agent_id },
-            payload: { subagent: finalized },
-          },
-          deps,
-          WORKBOARD_WS_AUDIENCE,
-        );
-      }
-
-      const result = WsSubagentCloseResult.parse({ subagent: finalized });
-      return { request_id: msg.request_id, type: msg.type, ok: true, result };
-    } catch (err) {
-      return workboardErrorResponse(msg.request_id, msg.type, err, deps);
-    }
-  }
+  const subagentRes = await handleSubagentMessage(client, msg, deps);
+  if (subagentRes) return subagentRes;
 
   const workboardRes = await handleWorkboardMessage(client, msg, deps);
   if (workboardRes) return workboardRes;
@@ -1553,362 +1110,8 @@ export async function handleClientMessage(
     return { request_id: msg.request_id, type: msg.type, ok: true, result };
   }
 
-  if (
-    msg.type === "memory.search" ||
-    msg.type === "memory.list" ||
-    msg.type === "memory.get" ||
-    msg.type === "memory.create" ||
-    msg.type === "memory.update" ||
-    msg.type === "memory.delete" ||
-    msg.type === "memory.forget" ||
-    msg.type === "memory.export"
-  ) {
-    if (client.role !== "client") {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unauthorized",
-        "only operator clients may access memory APIs",
-      );
-    }
-
-    if (!deps.memoryV1Dal) {
-      return errorResponse(
-        msg.request_id,
-        msg.type,
-        "unsupported_request",
-        "memory v1 not supported",
-      );
-    }
-
-    try {
-      if (msg.type === "memory.search") {
-        const parsedReq = WsMemorySearchRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const payload = parsedReq.data.payload;
-        const limit = Math.max(1, Math.min(500, payload.limit ?? 50));
-        const res = await deps.memoryV1Dal.search({ ...payload, limit });
-        const result = WsMemorySearchResult.parse(res);
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.list") {
-        const parsedReq = WsMemoryListRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const payload = parsedReq.data.payload;
-        const limit = Math.max(1, Math.min(500, payload.limit ?? 50));
-        const res = await deps.memoryV1Dal.list({
-          filter: payload.filter,
-          limit,
-          cursor: payload.cursor,
-        });
-        const result = WsMemoryListResult.parse({
-          v: 1,
-          items: res.items,
-          ...(res.next_cursor ? { next_cursor: res.next_cursor } : {}),
-        });
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.get") {
-        const parsedReq = WsMemoryGetRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const item = await deps.memoryV1Dal.getById(parsedReq.data.payload.memory_item_id);
-        if (!item) {
-          return errorResponse(msg.request_id, msg.type, "not_found", "memory item not found");
-        }
-
-        const result = WsMemoryGetResult.parse({ v: 1, item });
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.create") {
-        const parsedReq = WsMemoryCreateRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const item = await deps.memoryV1Dal.create(parsedReq.data.payload.item);
-        const result = WsMemoryCreateResult.parse({ v: 1, item });
-
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "memory.item.created",
-            occurred_at: item.created_at,
-            payload: { item },
-          },
-          deps,
-          OPERATOR_MEMORY_EVENT_AUDIENCE,
-        );
-
-        await maybeRunMemoryV1BudgetConsolidation({
-          deps,
-          op: "create",
-          agentId: item.agent_id,
-          memoryItemId: item.memory_item_id,
-        });
-
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.update") {
-        const parsedReq = WsMemoryUpdateRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const payload = parsedReq.data.payload;
-        const item = await deps.memoryV1Dal.update(payload.memory_item_id, payload.patch);
-        const result = WsMemoryUpdateResult.parse({ v: 1, item });
-
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "memory.item.updated",
-            occurred_at: item.updated_at ?? new Date().toISOString(),
-            payload: { item },
-          },
-          deps,
-          OPERATOR_MEMORY_EVENT_AUDIENCE,
-        );
-
-        await maybeRunMemoryV1BudgetConsolidation({
-          deps,
-          op: "update",
-          agentId: item.agent_id,
-          memoryItemId: item.memory_item_id,
-        });
-
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.delete") {
-        const parsedReq = WsMemoryDeleteRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const payload = parsedReq.data.payload;
-        const tombstone = await deps.memoryV1Dal.delete(
-          payload.memory_item_id,
-          { deleted_by: "operator", reason: payload.reason },
-          undefined,
-        );
-        const result = WsMemoryDeleteResult.parse({ v: 1, tombstone });
-
-        broadcastEvent(
-          {
-            event_id: crypto.randomUUID(),
-            type: "memory.item.deleted",
-            occurred_at: tombstone.deleted_at,
-            payload: { tombstone },
-          },
-          deps,
-          OPERATOR_MEMORY_EVENT_AUDIENCE,
-        );
-
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      if (msg.type === "memory.forget") {
-        const parsedReq = WsMemoryForgetRequest.safeParse(msg);
-        if (!parsedReq.success) {
-          return errorResponse(
-            msg.request_id,
-            msg.type,
-            "invalid_request",
-            parsedReq.error.message,
-            {
-              issues: parsedReq.error.issues,
-            },
-          );
-        }
-
-        const payload = parsedReq.data.payload;
-        const outcome = await deps.memoryV1Dal.forget({
-          selectors: payload.selectors,
-          deleted_by: "operator",
-        });
-        const result = WsMemoryForgetResult.parse({
-          v: 1,
-          deleted_count: outcome.deleted_count,
-          tombstones: outcome.tombstones,
-        });
-
-        for (const tombstone of outcome.tombstones) {
-          broadcastEvent(
-            {
-              event_id: crypto.randomUUID(),
-              type: "memory.item.forgotten",
-              occurred_at: tombstone.deleted_at,
-              payload: { tombstone },
-            },
-            deps,
-            OPERATOR_MEMORY_EVENT_AUDIENCE,
-          );
-        }
-
-        return { request_id: msg.request_id, type: msg.type, ok: true, result };
-      }
-
-      const parsedReq = WsMemoryExportRequest.safeParse(msg);
-      if (!parsedReq.success) {
-        return errorResponse(msg.request_id, msg.type, "invalid_request", parsedReq.error.message, {
-          issues: parsedReq.error.issues,
-        });
-      }
-
-      if (!deps.artifactStore) {
-        return errorResponse(
-          msg.request_id,
-          msg.type,
-          "unsupported_request",
-          "memory.export not supported",
-        );
-      }
-
-      const payload = parsedReq.data.payload;
-
-      const items: unknown[] = [];
-      let cursor: string | undefined;
-      for (;;) {
-        const page = await deps.memoryV1Dal.list({ filter: payload.filter, limit: 200, cursor });
-        items.push(...page.items);
-        cursor = page.next_cursor;
-        if (!cursor) break;
-      }
-
-      const tombstones: unknown[] = [];
-      if (payload.include_tombstones) {
-        let tCursor: string | undefined;
-        for (;;) {
-          const page = await deps.memoryV1Dal.listTombstones({ limit: 200, cursor: tCursor });
-          tombstones.push(...page.tombstones);
-          tCursor = page.next_cursor;
-          if (!tCursor) break;
-        }
-      }
-
-      const exportedAt = new Date().toISOString();
-      const exportArtifact = {
-        v: 1,
-        exported_at: exportedAt,
-        filter: payload.filter,
-        include_tombstones: payload.include_tombstones,
-        items,
-        ...(payload.include_tombstones ? { tombstones } : {}),
-      };
-
-      const ref = await deps.artifactStore.put({
-        kind: "file",
-        body: Buffer.from(JSON.stringify(exportArtifact, null, 2), "utf8"),
-        mime_type: "application/json",
-        labels: ["memory", "memory_v1", "export"],
-      });
-
-      const result = WsMemoryExportResult.parse({ v: 1, artifact_id: ref.artifact_id });
-
-      broadcastEvent(
-        {
-          event_id: crypto.randomUUID(),
-          type: "memory.export.completed",
-          occurred_at: exportedAt,
-          payload: { artifact_id: ref.artifact_id },
-        },
-        deps,
-        OPERATOR_MEMORY_EVENT_AUDIENCE,
-      );
-
-      return { request_id: msg.request_id, type: msg.type, ok: true, result };
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      const errorCode =
-        err &&
-        typeof err === "object" &&
-        "code" in err &&
-        typeof (err as { code?: unknown }).code === "string"
-          ? (err as { code: string }).code
-          : undefined;
-      if (message === "memory item not found") {
-        return errorResponse(msg.request_id, msg.type, "not_found", "memory item not found");
-      }
-      if (message === "invalid cursor") {
-        return errorResponse(msg.request_id, msg.type, "invalid_request", "invalid cursor");
-      }
-      if (message.startsWith("incompatible patch fields for kind=")) {
-        return errorResponse(msg.request_id, msg.type, "invalid_request", message);
-      }
-      if (errorCode === "invalid_request") {
-        return errorResponse(msg.request_id, msg.type, "invalid_request", message);
-      }
-      deps.logger?.error("ws.memory_request_failed", {
-        request_id: msg.request_id,
-        client_id: client.id,
-        request_type: msg.type,
-        error: message,
-        error_code: errorCode,
-      });
-      return errorResponse(msg.request_id, msg.type, "internal_error", message);
-    }
-  }
+  const memoryRes = await handleMemoryMessage(client, msg, deps);
+  if (memoryRes) return memoryRes;
 
   if (msg.type === "presence.beacon") {
     if (!deps.presenceDal || !client.device_id) {
